+++ app/src/main/AndroidManifest.xml
@@ -10,14 +10,14 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:theme="@style/Theme.Experimenting"
-        android:name=".application.ExperimentingApp"
+        android:theme="@style/Theme.XpImprover"
+        android:name=".application.XpImproverApp"
         tools:targetApi="31">
         <activity
             android:name=".MainActivity"
             android:exported="true"
             android:label="@string/app_name"
-            android:theme="@style/Theme.Experimenting">
+            android:theme="@style/Theme.XpImprover">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
