+++ app/src/main/java/com/pliniodev/xpimprover/resources/HomeStrings.kt
@@ -1,4 +1,4 @@
-package com.pliniodev.experimenting.resources
+package com.pliniodev.xpimprover.resources
 
 import cafe.adriel.lyricist.LyricistStrings
