+++ app/src/main/java/com/pliniodev/xpimprover/theme/Theme.kt
@@ -1,4 +1,4 @@
-package com.pliniodev.experimenting.theme
+package com.pliniodev.xpimprover.theme
 
 import android.app.Activity
 import android.os.Build
@@ -38,7 +38,7 @@ private val LightColorScheme = lightColorScheme(
 )
 
 @Composable
-fun ExperimentingTheme(
+fun XpImproverTheme(
     darkTheme: Boolean = isSystemInDarkTheme(),
     // Dynamic color is available on Android 12+
     dynamicColor: Boolean = true,
