+++ app/src/main/java/com/pliniodev/xpimprover/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.pliniodev.experimenting.theme
+package com.pliniodev.xpimprover.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
