+++ app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">Experimenting</string>
+    <string name="app_name">Mentor Xp Improver</string>
 </resources>
