+++ feature/src/main/java/com/pliniodev/feature/di/StepDi.kt
@@ -1,7 +1,7 @@
 package com.pliniodev.feature.di
 
 import cafe.adriel.voyager.core.registry.screenModule
-import com.pliniodev.experimenting.navigation.SharedScreen
+import com.pliniodev.xpimprover.navigation.SharedScreen
 import com.pliniodev.feature.FeatureHomeScreen
 
 val featureScreenModule = screenModule {
